NUMBER GUESSING GAME
# C++ Number Guessing Game:

Welcome to the Number Guessing Game implemented in C++! This is a simple console-based game where the computer selects a random number between 1 and 100, and the player has to guess the number. The game provides feedback on whether the guessed number is too high or too low until the player correctly guesses the number.